replace in-source packaging with out of source package building
Previously, the compiled binaries would be copied to the package
directory in the source directory in order to bundle the package
there and then copy the resulting archive to the build folder.

This comes with the problem that it generates binary artifacts in the
src directory that don't belong there.

Now, the required files for packaging are copied from the sources
dir into the build dir, binaries are copied to the copied packaging
files and stuff is bundled directly inside the build folder.

- out of source building
- the nupkg .NET version is automatically extracted from the nuspec.
  Define the .NET version in '<dependencies>':
  ```
  <dependencies>
      <group targetFramework="net6.0" />
  </dependencies>
  ```
- removed the '--clean' option. No clean needed since any
  artifacts are created in the build folder